web/requirements.txt
coverage==4.5.1
flake8==3.5.0
Flask==1.0.2
Flask-Migrate==2.2.0
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
marshmallow
Flask-WTF==0.14.2
psycopg2-binary==2.7.4
gunicorn
pytest
pytest-black
pytest-isort
pytest-flakes
factory_boy
PyJWT
